+++ lib/gorge/importer/round_importer.rb
@@ -12,6 +12,9 @@ module Gorge
       end
 
       def import
+        map_cache = generate_map_cache
+        location_cache = generate_location_cache(map_cache)
+
         round_count = round_info.count
 
         total_batch_count = (round_count.to_f / ROUND_IMPORT_BATCH_SIZE).ceil
@@ -22,9 +25,28 @@ module Gorge
           new_round_hashes, skipped_rounds_count = *args
 
           new_round_data = new_round_hashes.map do |hsh|
+            # startingLocation1 / startingLocation2 are *1-based* indices of
+            # the names in this array.
+            locations_ary = JSON.parse(hsh.fetch(:locationNames))
             [
               hsh.fetch(:roundId),
+              map_cache.fetch(hsh.fetch(:mapName)).id,
+              location_cache.fetch(
+                hsh.fetch(:mapName)
+              ).fetch(
+                locations_ary[
+                  hsh.fetch(:startingLocation2).to_i - 1
+                ]
+              ).id,
+              location_cache.fetch(
+                hsh.fetch(:mapName)
+              ).fetch(
+                locations_ary[
+                  hsh.fetch(:startingLocation1).to_i - 1
+                ]
+              ).id,
               hsh.fetch(:roundDate),
+              hsh.fetch(:roundLength),
               hsh.fetch(:maxPlayers1),
               hsh.fetch(:maxPlayers2),
               hsh.fetch(:tournamentMode) == 1,
@@ -36,7 +58,11 @@ module Gorge
           Round.import(
             [
               :round_id,
+              :map_id,
+              :alien_starting_location_id,
+              :marine_starting_location_id,
               :timestamp,
+              :length,
               :max_players_marines,
               :max_players_aliens,
               :tournament_mode,
@@ -81,6 +107,43 @@ module Gorge
 
       def round_info
         @source_db.from(:RoundInfo)
+      end
+
+      def generate_map_cache
+        map_count = Map.count
+        @l.debug({ msg: 'Generating round cache.' })
+
+        cache = @source_db.from(:RoundInfo).select_map { distinct(mapName) }.map do |name|
+          [name, Map.get_or_create(name)]
+        end.to_h
+
+        @l.debug({ msg: 'Map cache generated.', new_maps: Map.count - map_count })
+
+        cache
+      end
+
+      def generate_location_cache(map_cache)
+        location_count = Location.count
+        @l.debug({ msg: 'Generating location cache.' })
+
+        cache = {}
+
+        @source_db.from(:roundInfo).select { [locationNames, mapName] }.group_by(:mapName, :locationNames).each do |hsh|
+          map = map_cache.fetch(hsh.fetch(:mapName))
+
+          # { 'location_name': Location }
+          inner_hsh = JSON.parse(hsh[:locationNames]).map do |name|
+            [name, Location.get_or_create(name, map: map)]
+          end.to_h
+
+          if cache.key? map.name
+            inner_hsh = inner_hsh.merge(cache[map.name])
+          end
+
+          cache[map.name] = inner_hsh
+        end
+
+        cache
       end
     end
   end