+++ CHANGELOG.md
@@ -8,9 +8,16 @@ glance - what to expact from upgrading to a new version.
 
 ### Added
 
+- Importer: Store round lengths.
+- Importer: Store map names.
+- Importer: Store marine and alien starting locations.
+- Importer: Store time spent per player for each class (lifeform, weapon, ...)
+
 ### Changed
 
 ### Fixed
+
+- Attempt to create `DataImport::STORAGE_PATH` if it doesn't exist yet.
 
 ### Security